+++ projects/swimlane/ngx-ui/src/lib/components/json-editor/json-editor-flat/json-editor-node-flat/node-types/object-node-flat/object-node-flat.component.ts
@@ -43,6 +43,8 @@ export class ObjectNodeFlatComponent extends ObjectNode implements OnInit, OnCha
 
   @Input() isDuplicated = false;
 
+  @Input() passwordToggleEnabled = false;
+
   indentationArray: number[] = [];
 
   duplicatedFields = new Map<string, string>();