+++ projects/swimlane/ngx-ui/src/lib/components/json-editor/json-editor/json-editor-node/node-types/array-node/array-node.component.ts
@@ -8,4 +8,6 @@ import { ArrayNode } from '../../../../node-types/array-node.component';
 })
 export class ArrayNodeComponent extends ArrayNode {
   @Input() isDuplicated = false;
+
+  @Input() passwordToggleEnabled = false;
 }