+++ projects/swimlane/ngx-ui/src/lib/components/json-editor/json-editor/json-editor.component.ts
@@ -26,6 +26,8 @@ export class JsonEditorComponent extends JsonEditor {
 
   @Input() typeCheckOverrides?: any;
 
+  @Input() passwordToggleEnabled = false;
+
   @ContentChildren(JsonEditorNodeComponent)
   nodeElms: QueryList<JsonEditorNodeComponent>;